Move inner config classes to own files
