+++ src/main/java/net/smartcosmos/cluster/auth/config/AuthServerConfiguration.java
@@ -6,6 +6,8 @@ import org.springframework.cache.concurrent.ConcurrentMapCache;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.core.userdetails.UserCache;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 
 import net.smartcosmos.cluster.auth.SmartCosmosUserCache;
 
@@ -26,5 +28,11 @@ public class AuthServerConfiguration {
     Cache simpleUserCache() {
 
         return new ConcurrentMapCache("SmartCosmosUserCache", false);
+    }
+
+    @Bean
+    PasswordEncoder passwordEncoder() {
+
+        return new BCryptPasswordEncoder();
     }
 }
